chore: Split the database structure internal types
