+++ meilidb-data/src/database/raw_index.rs
@@ -1,0 +1,9 @@
+use super::{MainIndex, WordsIndex, DocsWordsIndex, DocumentsIndex};
+
+#[derive(Clone)]
+pub struct RawIndex {
+    pub main: MainIndex,
+    pub words: WordsIndex,
+    pub docs_words: DocsWordsIndex,
+    pub documents: DocumentsIndex,
+}
